Update hangar-frontend-connection.php
Pulled my hair out to know why my translation of this plugin didn't work. But as it turns out, your function is only defined, never actually executed. I don't know if this is the correct place to call it, but at least now it works for me :)

Best regards,
Haakon Sundaune
Oslo, Norway